pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neutron-skills"
version = "0.1.0"
description = "A curated registry of Agent Skills for neutron scattering."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "BSD-3-Clause"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["neutron", "scattering", "agent-skills", "llm", "agents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]

# Core dependencies
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Development tools
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

# Web application dependencies
web = [
    "flask>=3.0.0",
    "jinja2>=3.1.0",
    "bootstrap-flask>=2.3.0",
]

# API/MCP server dependencies
api = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
]

# CLI extras (click is already a core dep; this extra adds progress bars & env loading)
cli = [
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
]

# Data science extras
science = [
    "scipy>=1.11.0",
    "matplotlib>=3.8.0",
    "plotly>=5.17.0",
    "scikit-learn>=1.3.0",
]

# Documentation
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]

# All optional dependencies
all = [
    "neutron-skills[dev,web,api,cli,science,docs]",
]

[project.urls]
Homepage = "https://github.com/yourusername/your-repo"
Documentation = "https://github.com/yourusername/your-repo#readme"
Repository = "https://github.com/yourusername/your-repo"
Issues = "https://github.com/yourusername/your-repo/issues"

[project.scripts]
neutron-skills = "neutron_skills.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
neutron_skills = ["py.typed", "skills/**/*"]

# Pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",                      # Show summary of all test outcomes
    "--strict-markers",         # Error on unknown markers
    "--strict-config",          # Error on unknown config options
    "--cov=src/neutron_skills", # Coverage for your package
    "--cov-report=term-missing", # Show missing lines in coverage
    "--cov-report=html",        # Generate HTML coverage report
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# Ruff configuration (linting and formatting)
[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex (let developers decide)
    "W191",   # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
"tests/**/*.py" = [
    "ARG",    # Unused function arguments in tests are OK (fixtures)
    "S101",   # Assert statements are OK in tests
]

[tool.ruff.lint.isort]
known-first-party = ["neutron_skills"]

# Black configuration (code formatting)
[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

# MyPy configuration (type checking)
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Set to true for stricter type checking
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
